Backend/requirements.txt
alembic==1.14.1
aniso8601==10.0.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Migrate==4.1.0
Flask-MySQLdb==2.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
mysql-connector==2.2.9
mysql-connector-python==9.1.0
mysql-connector-python-rf==2.2.2
mysqlclient==2.2.7
numpy==2.2.2
packaging==24.2
pandas==2.2.3
psycopg==3.2.4
psycopg2==2.9.10
pycparser==2.22
pygame==2.6.1
PyJWT==2.10.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
WTForms==3.2.1
